[Feat] [FE-16] BlackBadge 임의의 텍스트를 받을 수 있도록 변환
